Reflect review comment
